+++ Plugin/UI/Constants/Fonts.cs
@@ -4,36 +4,39 @@ namespace Silksong.BingoSync.UI.Constants;
 
 internal static class Fonts
 {
-	private static readonly Lazy<(Font? Bold, Font? Normal)> TrajanFonts = new(() =>
+	private static readonly Dictionary<string, Font> CachedFonts = new();
+
+	/// <summary>
+	/// Gets the <see cref="Font"/> with the given name
+	/// </summary>
+	private static Font? GetFont(string name)
 	{
-		Font? normal = null;
-		Font? bold = null;
+		if (CachedFonts.TryGetValue(name, out var font))
+			return font;
 
 		foreach (var f in Resources.FindObjectsOfTypeAll<Font>())
-		{
-			if (f == null)
-				continue;
+			CachedFonts[f.name] = f;
 
-			switch (f.name)
-			{
-				case "TrajanPro-Bold":
-					bold = f;
-					break;
-				case "TrajanPro-Regular":
-					normal = f;
-					break;
-			}
-		}
-		return (bold, normal);
-	});
+		return CachedFonts.GetValueOrDefault(name);
+	}
 
 	/// <summary>
-	/// Accesses to the normal font
+	/// Gets the normal font
 	/// </summary>
-	public static Font? Normal => TrajanFonts.Value.Normal;
+	public static Font? Normal => GetFont("TrajanPro-Regular");
 
 	/// <summary>
-	/// Accesses to the bold font
+	/// Gets the bold font
 	/// </summary>
-	public static Font? Bold => TrajanFonts.Value.Bold;
+	public static Font? Bold => GetFont("TrajanPro-Bold");
+
+	/// <summary>
+	/// Gets the 'Arial' font
+	/// </summary>
+	public static Font? Arial => GetFont("ARIAL");
+
+	/// <summary>
+	/// Gets the default <see cref="Font"/>
+	/// </summary>
+	public static Font? Default => Font.GetDefault();
 }