+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 <summary>v1.2.2</summary>
 
 ## Additions
+- Added config `BoardCellFont` that allows to change the font of the board's cells for better readability
 - Added condition `has_obtained_quill` with 3 simple goals
 - Added condition `has_obtained_map` with 28 simple goals
 - Added condition `has_completed_wish` with 54 simple goals